+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v1.2.0 (Upcomming)
 
+- Added safely.GoDo(cxt, GoDoer)
+* NewReroute(route string) can be used to create Reroutes now.
 * From() now takes an vararg of strings: `From("cxt:foo", "cxt:bar")`.
   It can still take a space-delimited set of strings, too. In fact, both
   can be used together:`From("cxt:foo", "cxt:bar cxt:baz")`.